Move fflags to separate module
